DecAgg-v2.4 (1) Fix the data receive and aggregation bugs. (2) Code modularity.
BUG TO FIX:
Check the log thoroughly and find if there are any unexpected behaviors.
(1) Initial interest managment.
Initial interest generated by producers should not be processed by the strategy. Instead, it should just directly be sent to the aggregate or the cloest gate forwarder in the network. Now, the log shows 'OPTIMIZATION: All 2 IDs route to the same face (ID: 257).'
(2) PIT interest deletion.
Once the data arrive at the aggregator and match the PIT entry, the PIT entry should be deleted. Now, the log shows that the PIT entry is not deleted, and still take up for the memory.
For example, see log of '!!! C1 RECEIVED DATA ON FACE 257: /aggregate/%01/seq=0'. At this time, previous satisfied interest should be deleted.